Distribute work over cluster members
A small experiment we cooked up at the 2017-02-16 Amsterdam.scala meetup
(https://www.meetup.com/amsterdam-scala/events/237506661/).

It adds a WorkCreatorActor which periodically creates work and uses a
Router to distribute it among workers in a round-robin fashion. Whenever
a new cluster member joins, akka remoting is used to dynamically create
a number of workers on the joined node and adds them to the worker pool.

In a real project you'd be more likely to use higher-level concepts
(ClusterSingleton, perhaps Cluster Sharding and/or some kind of dedicated work
queue), but it was a fun way to explore some lesser-known parts of Akka.